removing defects
