+++ pymatgen/transformations/tests/test_defect_transformations.py
@@ -3,7 +3,7 @@
 
 import unittest
 
-from pymatgen.analysis.defects.core import Vacancy
+from pymatgen.analysis.defects_old.core import Vacancy
 from pymatgen.transformations.defect_transformations import DefectTransformation
 from pymatgen.util.testing import PymatgenTest
